Module type Template.RENDERABLE

The RENDERABLE module type defines an interface for types that can be converted to key-value pairs for use in templating.

type t

val to_key_value_pairs : t -> (string * string) list

to_key_value_pairs t converts t to a list of key-value pairs.